# tests/fixtures/openapi-fixture.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hand-made test fixture, not CIPP",
    "version": "test"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "parameters": {
      "tenantFilter": {
        "name": "tenantFilter",
        "in": "query",
        "required": true,
        "description": "Target tenant",
        "schema": {
          "type": "string"
        }
      }
    }
  },
  "paths": {
...
    "/api/ExecCIPPDBCache": {
      "post": {
        "summary": "ExecCIPPDBCache",
        "operationId": "ExecCIPPDBCache",
        "tags": [
          "CIPP > Core"
        ],
        "description": "ExecCIPPDBCache description",
        "x-cipp-role": "CIPP.Core.ReadWrite",
        "parameters": [
          {
            "name": "Name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/tenantFilter"
          },
          {
            "name": "Types",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            }
          }
        }
      }
    },
...
  }
}